Refactor calendar screen
