+++ app/src/main/java/com/tpp/theperiodpurse/ui/calendar/components/MonthHeader.kt
@@ -1,0 +1,60 @@
+package com.tpp.theperiodpurse.ui.calendar.components
+
+import android.os.Build
+import androidx.annotation.RequiresApi
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.kizitonwose.calendar.core.CalendarMonth
+import com.tpp.theperiodpurse.ui.calendar.displayText
+import com.tpp.theperiodpurse.ui.onboarding.scaledSp
+import java.time.DayOfWeek
+import java.time.format.TextStyle
+import java.util.*
+
+@RequiresApi(Build.VERSION_CODES.O)
+@Composable
+fun MonthHeader(calendarMonth: CalendarMonth) {
+    // The header that is displayed for every month; contains week & month.
+    val daysOfWeek = calendarMonth.weekDays.first().map { it.date.dayOfWeek }
+    Column(
+        modifier = Modifier
+            .padding(vertical = 18.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        MonthAndDayHeadings(calendarMonth, daysOfWeek)
+    }
+}
+
+@RequiresApi(Build.VERSION_CODES.O)
+@Composable
+fun MonthAndDayHeadings(calendarMonth: CalendarMonth, daysOfWeek: List<DayOfWeek>) {
+    // Month
+    Text(
+        modifier = Modifier.padding(12.dp),
+        textAlign = TextAlign.Center,
+        fontSize = 18.scaledSp(),
+        fontWeight = FontWeight.W500,
+        text = calendarMonth.yearMonth.displayText()
+    )
+
+    // Days of Week
+    Row(modifier = Modifier.fillMaxWidth()) {
+        for (dayOfWeek in daysOfWeek) {
+            Text(
+                modifier = Modifier
+                    .weight(1f)
+                    .offset(y = 10.dp),
+                textAlign = TextAlign.Center,
+                fontSize = 10.scaledSp(),
+                fontWeight = FontWeight.ExtraBold,
+                text = dayOfWeek.getDisplayName(TextStyle.SHORT, Locale.getDefault())
+            )
+        }
+    }
+}
